requirements.txt
ccxt
xgboost
pandas
matplotlib
flask
python-dotenv
